replaced files with Blobs
Using static media files does not work because it exceeds the
file size limit on Google App Engine.

The work-around is to store the media files as Blobs in a database
and then serve parts of the blob on request. This requires creating
an upload service to be able to put blobs in to the store.