refactor csrf to avoid origin IP check
The blob upload implementation on GAE does not maintain the original
source IP address when passing control to the class that handles the
POST message. This caused the csrf checks to fail, because it was
using geo-ip information to check that the client that requested the
form was the same one that submitted the form.

The work-around is to use the Origin and Referer fields, which are
preserved.