use a fixed availabilityStartTime
Previously this service would dynamically select the
availabilityStartTime based upon the first request from a client,
so that the pseudo-live stream always starts at the beginning of
the stream. This does not reflect how live streams are used and also
the DASH specification mandates that the availabilityStartTime is a fixed
value.

This commit changes availabilityStartTime to be fixed at midnight
today (i.e. the day when the manifest is requested) and adjusts
the data in the segments to match this timeline.