add support for UTCTiming
This commit adds a /time endpoint that will respond with the server's
current time of day, in one of three formats:

    /time/xsd    XSDateTime
    /time/iso    ISO 8601
    /time/ntp    NTP timestamp format

This can be comined with a "drift" CGI parameter to give the server
a fake clock drift. For example:

    /dash/manifest_e.mpd?drift=30

will produce a manifest with a clock drift that is 30 seconds behind the
actual time of day.