fix audio/video sync
The audio and video synchronisation was completely incorrect when
operating in live mode.

The problem was that the audio and video representations needed to be
mapped back to a common origin time (the point where the stream last
looped). From this origin time, it is possible to work out the delta
between the requested time and the origin time. Finally this can be
mapped to a segment number.

The origin time then needs to be added to the base_media_decode_time
so that all representations start from the same origin time.