include duration when loading pre-generated index
When gen_index.py creates the index files, it puts the duration of
each segment in the file. However when segment.py is called to
load one of these generated files, it was not passing the duration
to the Box constructor, causing the duration information about each
segment to be lost.