add ability to inject errors into media responses
To allow testing the handling of HTTP errors, add some CGI parameters
that can be used to inject errors into HTTP responses.
The keys are:
    v404, v410, v503, v504, a404, a410, a503, a504

Each key accepts a comma separated list of times, in the form:
    HH:MM:SSZ

These times define a time of day that will return an error. Note that
this time refers to the time of day associated with the fragment,
not the time of day when the request is made. For example, the CGI
parameter v404=10:00:00Z will cause a 404 error when requesting a
video fragment that corresponds to 10am, regardless of the time of day
when the HTTP request is made.

When injecting a 503 or 504 error, only one error is returned per-segment,
per-client. If a client retries a request in response to a 503 or 504
error, the second attempt will succeed.