don't include an empty error injection CGI parameter
If the result of calculate_injected_error_segments() is an empty string
(e.g. because the times specified are too far in the past) the
calculate_dash_params() function was adding an empty CGI parameter. For
example "foo.mpd?404=". This causes a 400 error when this URL is used.

The solution is to only include the CGI parameter when there is at least
one value for that parameter.