add "start" parameter to control availabilityStartTime
By default availabilityStartTime is set to be the start of the
day. In some cases, it is useful to be able to control this
value. This commit adds a "start" parameter that can either contain
the string "now", which means use current time of day, or an
ISO datetime string (YYYY-MM-DDTHH:MM:SSZ).