add support for range requests of fragments
When using trick play, the DASH client might make HTTP range
requests. This commit adds support for returning ranges of a
fragment. This implementation will load the entire fragment into
memory and then slice out the required range. Ideally it would
only load the parts it needed and only update the fragment if
the range request covers the MOOF box.