fix "off by one" error in range responses
The implementation of HTTP range requests was incorrect, as the
end value is an inclusive value, whereas in Python the slice
operator excludes the end value. This meant that every HTTP range
request returned one byte less than it should have.

This commit also adds support for either the start or end value to
be omitted in the range request.