add support for W3C clearkey
This commit adds support for the W3C clear-key content protection
system. This system puts a URL in the manifest that can be used
to request the keys. Typically this URL is an HTTPS url. The
request is a POST with a json document that contains the key IDs.
The response is also a JSON document, that contains the keys for
each requested key ID.

The contents of both the request and response use a variation on
the normal base64 encoding by using different values for the
"+" and "/" character and by removing the "=" padding characters.

https://www.w3.org/TR/encrypted-media/#clear-key