tidy up layout of main page and add button to play video
To make it easier to tell whether the manifest or options is
active, but key labels on each column, which only appear when
the column is active.

Add a play button at the top right of the page that will play
the currently selected manifest using the currently selected
options.