use cursor based design for main page
To make it easier to interact with the main page, use a cursor
metaphor where the left, right, up and down keys move around
the grid. ENTER will select the manifest (if in the manifest
column), cycle between options (if in the option column) or play
the video (if in the "play video" cell).