add option to select between native and MsE/EME playback
Some browsers are capable of native DASH playback. This commmit adds
an option to allow switching between the native DASH client and
ShakaPlayer (which uses EME and MSE).

It also provides an option for encrypted streams to provide the
keys in the Shaka configuration, removing any need for a DRM CDM.
This allows checking if the stream is encrypted correctly.