store list of streams and media indexes in the datastore
Having to modify Python files in the media directory and then
re-deploy the service was becoming painful to maintain.

This commit removes all of the hard coded media information and
replaces it with entries in the DataStore. There is a model for
available streams and the MediaFile model has been updated to
add a JSON field that stores the index of that media file.

A new buffered IO class has been added, so that the mp4 parser
can rapidly index a media file that is stored in the Blobstore.
This was required because using the BlobReader directly is very
slow and does not implement the peek() function.