move PSSH boxes to the end of the MOOV
It appears that some devices expect the PSSH boxes to be at the
end of the MOOV, even though the ordering is not defined in the
MPEG DASH common encryption specification. However the DECE
common file format specification states that PSSH boxes are
placed after the trak and mvex boxes.

This commit moves any PSSH boxes that are injected into the
init segments to the end of the MOOV box.