re-enable video error injection support
The server supports injecting HTTP errors into the response at
selected points and also injecting invalid data into the compressed
video stream. This is used for testing how the DASH client and the
decoder cope with errors.

This commit fixes bugs that were introduced in this feature as
part of the refactoring for DRM support.