fix calculations of live segment values
The media time calculations for live streams was incorrect and
was generating incorrect values after the stream has looped.
Also, the use of floating point numbers to calculate position causes
rounding errors when there is several hours difference between
now and availabilityStartTime.

This commit changes the logic to use integer maths to calculate
position and handles the special case whereby the difference
between the duration of the audio and video representations can
cause one of them to loop one fragment before the other.