select reference representation from the current stream
When selecting the reference representation (the one that is used to
calculate positions in the live stream) it needs to pick one that
is one of the selected stream, rather than just picking the first
representation found in the database.

This bug shows up when there is more than one stream available, the
streams are of different durations and live stream mode is selected.
In this case the server will incorrectly return the first segment.

This commit fixes this bug so that the stream prefix is always used
when selecting the reference representation.