store Marlin and PlayReady license URLs in the streams table
Rather than having one hard coded SAS URL, make it a property of
the stream. At the same time, also add a field the for PlayReady
license URL. These fields are optional. If the PlayReady LA_URL
is not specified for a stream, the Microsoft test server is used.

The populate-db.py has also been updated to include these URLs
when adding a stream.