remove use of peek to parse MP4 box and descriptor headers
The peek() function is not guaranteed to return the number of
requested bytes. This means that it is not safe to use this function
for parsing the header of ISOBMFF boxes or descriptors, because
the call to peek() might not return sufficient data.

This commit refactors the MP4 and descriptor parser to use the
conventional read() method and then pass this parsed data
(as "initial_data") when the rest of the box is parsed to avoid
having to read and parse the same data twice.