add JSON output format to mp4 parser
To provide a more human readable version of the output from the
"--show" option, add support for JSON output.

This commit largely removes the use of the _field_repr() function in
the various atom classes, replacing then with a _to_json() function.
The responsibility of a _to_json() function is to copy all of the
fields and children of an atom into a new dictionary object. The
fields can mostly be kept in their original form, with the exception
of binary data, which needs to be wrapped inside a Binary class.

The public toJSON() and __repr__() APIs internally use these
_to_json() methods to make a nested dictionary and then translate
each field as required by the output format.