change main_audio CGI parameter to support a stream ID
If there are multiple audio streams that use the same codec,
the main_audio CGI field does not provide a way to select
which audio stream should be declared the main role.

This commit modifies main_audio to take either a codec name
or a stream ID as a way to indicate which stream is to be
the main one.