add playready_version parameter and v1 PlayReady UUID
In some situations it is useful to be able to limit the PlayReady
features to be compatible with a particular version of PlayReady.
This commit adds a "playready_version" CGI parameter that is
used to set the version parameter in the PlayReady class. It must
be set to a floating point number in the range 1.0 to 4.0.

When playready_version is set to 1.0 the schemeIdUri field in the
ContentProtection element will be set to the PlayReady v1.0 UUID
rather than the normal PlayReady UUID.