add inband and inline DASH event support to views
This commit adds support for DASH events in a manifest (inline
events) or in media segments (inband events).

The format for the DASH events is controlled by the `events`
CGI parameter, for example "events=ping".

One example DASH format is added in this commit "PingPong" which
simply alternates between payloads of "ping" and then "pong".

There are 8 parameters that can be used to modify the events:

* count - number of events to insert (0=unbounded for inline events)
* duration - duration of event (in timescale units)
* inband - boolean to select between inline and inband events
* interval - time between events (in timescale units)
* start - time when first event occurs (in timescale units)
* timescale - ticks per second for the event timeline
* value - the value string to use
* version - selects v0 or v1 DASH emsg format for inband events