only test options used by a manifest in exhaustive tests
There are some CGI options that are not used by a particular
manifest. It makes no sense to run any checks that modify those
unused CGI parameters.

This commit refactors the code to provide a "features" property
for each manifest that is a set that defines the CGI options
that can modify the manifest.

Another change is to provide a "restrictions" dictionary, that
defines which CGI options have a restricted set of allowed
options.

The final change in this commit is to split the test of each
manifest into its own manifest_*_test.py file. This made it
easier to test this refactorying and should help with future
changes to run these tests in parallel.

These changes reduce the test time on the development machine
from over 35 minutes to just over one minute.