start adding support for multi-period DASH streams
This commit starts the code changes for adding multi-period
support by creating a periods array in the template context
and updating hand_made.mpd, manifest_n.mpd and manifest_vod_aiv.mpd
to make use of this new template context.

A manifest has to list the "periods" feature for this new behaviour
to be enabled.