keep track of sample_count in saiz box
The saiz box contains a field that describes how many samples
that it provides size for. When a default_sample_info_size field
is provided, there is no need to list the size of each sample,
but the sample_count field is still required.

The saiz parser was incorrectly assuming that the sample_count field
was not relevant when the default_sample_info_size field is set.