allow query of contentType, prefix and encryption of media files
To save having to load and parse every JSON representation file when
filtering MediaFile entries, add contentType and encrypted fields to
MediaFile and use those to filter queries. Bitrate has also been
added to MediaFile, which is used to sort the results so that the
ordering of results is now consistent.

The MediaFile.all() function is also extended to support a query by
filename prefix, but this is handled within the class, rather than
using the NDB query system.