support senc box before the saiz box
The CencSampleEncryptionBox might occur in a stream before the
SampleAuxiliaryInformationOffsetsBox, however is it not possible
to parse the senc box until the size information the saiz box
has been parsed.

This commit adds a REQUIRED_PEERS class member varible, which
if no-none, contains a list of atom names. The parsing of a box
will be deferred until all of the REQUIRED_PEERS atoms have been
parsed.