pin google-cloud-sdk versions
The latest version of the Google cloud SDK causes the docker
container to fail to perform any ndb queries, with an error:

app must not be empty

The suggested work-arounds of setting APPLICATION_ID do not appear
to work.

As the Python2 version of GAE is at end-of-life, it is not worth
spending a lot of time trying to fix the issue. It is easier to
simply fix the GAE version.