add support for direct, ntp and sntp UTCTiming methods
This commit adds support for all of the UTCTiming element methods
described in the 2019 DASH specification. The UTC timing server
provided in this repository supports:

* HTTP HEAD
* HTTP xsd:dateTime
* HTTP ISO DateTIme
* HTTP NTP

If npt or sntp format is used, the Google public NTP time servers
are selected. The "time_value" CGI parameter can be used to replace
this with a different list of NTP time servers.