add option to dash_validator to set maximum duration of validation
If a stream has quite a long duration, it can take a long time to
validate the entire stream. This commit adds a "--duration" flag that
limits the number of segments that are checked. It also works when the
"--save" option is used, to limit how much of the stream is saved.