allow storing saved streams into single files for each representation
Add a new storage option for saving a DASH stream that matches the
form used for uploading media to the server. It stores each
representation in its own file, with a common prefix. In addittion,
a JSON file is generated that can be used by populate-db.sh to
upload the resulting stream.