update trun.data_offset field if the size of the moof changes
When modifying the moof, the data_offset field in the trun might need
to be updated, as it points to the sample data in the MDAT box. If
the size of the moof changes, these pointers need to be updated.

This commit makes the assumption that the mdat box will always
directly follow the moof box, and that the mdat box will only have
a simple 8 byte atom header.