enable creating streams with bugs in them
Sometimes it is useful to be able to generate a DASH stream that
contains known errors. This might be for checking that the DASH
validator detects this issue, or it might be to check a DASH client
to see how it handles such a stream.

This commit adds a "bugs" CGI parameter that can be used to tell
the server to deliberately include the named error in the stream.
At the moment, the only known error is "saio" which causes the
MP4 generator to produce incorrect SAIO offset values.