add support for H265 representations
This commit adds support for h.265 (HEVC) fragments that have
been encoded using the "hev1" or "hvc1" schemes.

Details of the correct codec tag for HEVC is defined in
Annex E of ISO/IEC 14496-15. However, the details of how to signal the
general_profile_compatibility_flags field is not very clear. From
experimentation looking at existing streams from DASH-IF [1] the field
appears to be least significant bit first, where the first bit in this
field represents the first profile.

[1] https://dashif.org/codecs/video/

See also:
* https://github.com/paulhiggs/codec-string/blob/main/decode-hevc.js
* https://chromium.googlesource.com/chromium/src/media/+/master/base/video_codecs.cc
* https://tools.axinom.com/capabilities/media?cd=h265&h265-se=hev1,hev2,hvc1,hvc2&h265-p=m
* https://github.com/cloudwebrtc/ireader-media-server/blob/master/libflv/source/mpeg4-hevc.c