sample_composition_time_offset field in the trun can be a signed int
Depending upon the value of the "version" field in the trun FullBox,
the sample_composition_time_offset might be a signed or an
unsigned int.

The parser function for the "trun" was using signed int and the
encoder was using unsigned! This commit fixes this so that they both
check the version field.