add support for hvc1 sample entry format
Section 8.3.2 of ISO/IEC FDIS 14496-15:2019(E) states:

Parameter sets: A parameter set to be used in a picture must be sent
prior to the sample containing that picture or in the sample for that
picture. For a video stream that a particular sample entry applies to,
the video parameter set, sequence parameter sets, and picture parameter
sets, shall be stored only in the sample entry when the sample entry
name is ‘hvc1’, and may be stored in the sample entry and the samples
when the sample entry name is ‘hev1’.

The "hev1" sample entry format was already supported. This commit adds
the "hvc1" sample entry format.