refactor generating input form fields into each model
Rather than have the request handlers create the fields
array used to generate the HTML form for adding or
editing a model, put this logic in the model.

This has the advantage of putting the output of the field
near to the definition of the field in the model. This reduces
the chances of them getting out of sync.