refresh CSRF token if has been used and not replaced
If a call to an API fails, the response might not contain a new
CSRF token for the next API call. To avoid accidental re-use of
a CSRF token, set it to None when used and refresh the token
if an API call is made when token is still None.