port database download script to new codebase
Convert the "download-db.py" script from the Python 2.7 codebase to
the new codebase. There is a change in functionality during this
process whereby each stream is saved into its own directory, with
its own JSON file.

The original "downloaded.json" file is preserved, that contains all of
the data from all of the streams.