use nginx and uwsgi to serve HTTP requests
Rather than use the Flask development test server to serve HTTP
requests, use nginx as the HTTP server and uwsgi to run the
Python app.

It is still possible to run the Flask development test server
by either setting the docker build target to "base" or by setting
the entrypoint when running the Docker container.