rewrite how live segment time and positions are calculated
The code that calculated the segment numbers and display time for
each segment wasn't working correctly with large differences between
availabilityStartTime and current time. The symptom of this problem
was when DASH clients requested audio segments, the server would
respond with a 404 error even though the fragment should exist.

One issue was the use of floating point numbers was causing a loss
of precision. A second problem was that the difference in duration
between the reference representation and the representation being
served was causing audio and video adaptation sets to drift out of
sync.