use track ID and start number from mediafiles
When choosing the "id" attribute for each AdaptationSet, use the
track ID from the representation mediafile's track ID. This allows
multiple audio files to be uploaded for a stream that can either be
individual tracks or different bitrates for one track.

The startNumber attribute is now taken from the segment number
field in the representation's mediafile. This means that it provides
correct segment numbering for VOD streams, even if the mediafile
doesn't start at segment 1.