use decorators to store mp4 atom and descriptor tables
Rather than store the dictionaries of mp4 atoms and mp4 descriptors
within the Mp4Atom and Descriptor classes, define decorator functions
that can applied to class definitions.

It makes it simpler to define each mp4 box and its fourcc code as
it just requires an @fourcc() decorator applied to the class the
implements the atom.

It also paves the way for being able to separate out these boxes
into their own files, to make the mp4 parsing code to be a bit
easier to maintain.