support $Time$ for both vod and live streams
Yet another rewrite of segment timeline calculations! This time, the
algorithm is to map the duration of the reference media to the
timeline of the representation. This value is then used to calculate
the number of times the stream has looped. From there, a simple linear
search is used to find to the next segment starting after the target.