use separate sessions for each WebSocket validation session
If more than one browser was connected the validator, each would
receive the same output.

This commit puts all of the state information for a websock connection
into an instance of ClientConnection class. This instance is used to
respond to the client that opened the connection and no other client.