use a different home page for browsers that support modules
When running on a modern browser, show a home page that re-uses
the settings from stream defaults to allow the user to configure
every option. To avoid overloading the user, by default only the
featured options are shown. A "show all options" switch is
provided that shows all available options.