use all periods when deciding if validation is complete
When validating a multi-period DASH stream, the validator was
treating each period individually when applying duration limits.
It should be starting at the first period and using each period's
duration. For example if validation duration was set to 25 seconds
and there are 4 periods of 10 seconds each, it should be using the
first three periods, with just 5 seconds of the third period.