add a FLASK_DASH__PROXY_DEPTH to control number of proxy hosts to trust
When using a reverse proxy, or many Python hosting platforms, the
proxy will intercept and forward all external requests to the
local WSGI server.

From the WSGI server and Flask application’s perspectives, requests
are now coming from the HTTP server to the local address, rather
than from the remote address to the external server address.

The FLASK_DASH__PROXY_DEPTH setting is used to control how many proxy
servers to trust. When running inside a Docker container, this value
must be at least one, as the WSGI server inside the container will be
proxied by an nginx server that is also running inside the container.