check upstream X-Forwarded-Proto headers
When nginx sets the X-Forwarded-Proto header on the requests it
sends to the WSGI server, it needs to check if there is another
reverse-proxy upstream of itself, which might have set the
X-Forwarded-Proto header.

If the X-Forwarded-Proto is present in the incoming request, it
is passed onto the WSGI server. If not present, the HTTP scheme
of the request to nginx is used.