dont try to Alembic upgrade a new database
When a new database is created, it will use the latest database
Schema. In this situation, Alembic must not try to upgrade the db
when it is called, as the database is already upto date.

This commit sets the Alembic version information to be the current
head version on the second time the server is started with a new
database.