add support for MPD patches for live streams
The 2022 version of the DASH specification added support for
MPD patches. These are XML documents that can be applied to an
in-memory version of a complete MPD to modify the manifest. Its
primary use is to avoid having to keep fetching the whole manifest
for live streams that use SegmentTimeline.

This initial implementation just replaces the SegmentTimeline elements
in each AdaptationSet. Ideally this will be extended to use the "add"
patch command to just add the new <S> nodes to the existing in-memory
SegmentTimeline.