refactor mock time to allow tests to call sleep functions
In some situations, the code under test needs to be able to call
time.sleep() or asyncio.sleep(). We don't want to have to wait for
real time to elapse and also some of these tests start from a fixed
mock time-of-day.

This commit refactors the time mocking to support mocking getting the
current time of day from both the datetime and time packages, and
mocking the sleep functions in the time and asyncio packages.