dont recreate media segment list during media prefetch
When prefetch_media_info() is called prior to validation, it
was causing the media segment list in each representation to
be re-created. This is a problem when validating a stream where
the validation duration is greater than timeshift buffer depth.

In that scenario, the validator needs multiple refreshes of the
manifest to have validated a sufficient duration.