move manifest context generation into its own class
This commit removes the logic from the request handler base class
that generated the context used for servicing DASH requests. The
code is moved into a new ManifestContext class.

This simplifies the request hander base class and allows type hints
to be defined for the data passed into the templates for rendering
manifest templates.