add command line utility to reset a user's password
if the admin password has been lost, there's no way to log into the
server without manually changing the password using sqlite. This isn't
easy as the password is salted and hashed.

This commit adds a command line util to allow a password to be reset.
It will set the password and also set the "must change" flag to remind
the admin that they need to change it, as the current value will now
probably be in someone's bash history.