list each media file in stream edit page
commit 48b7072329696c192265f298a0e9d774ce7e3f77 introduced a regression
that caused the stream edit page to no longer show the list of media
files for a given stream. This was not detected by unit tests as
this table was not being checked.

This commit fixes the regression and adds additional tests to hopefully
detect if there are any future regressions.