enforce that track IDs are unique within a single stream
to allow the IDs assigned to adaptation sets to be simplified, it is
useful to enforce a rule that within a single stream track IDs are only
shared by media files that are part of a single AdaptationSet.

For example all video files within a stream can use track ID one, but
the audio files for that stream must use a different track ID.

This commit adds an Alembic migration that will check this rule and
if required create new media files that change the track IDs.