timescale in media fragments doesn't have to match timescale in manifest
The DASH specification does not mandate that the timescale found
in MP4 fragments has the match the timescale defined in the
AdaptationSet.

This commit checks for the special case and adjusts the times and
durations it expects to allow for the difference between timescales.
The validator mandates that AdaptationSet timescale and the media
timescale are multiples of each other. The DASH specification doesn't
mandate this. It states for the @timescale attribute:

    NOTE: This can be any frequency but typically is the media clock
    frequency of one of the media streams (or a positive integer
    multiple thereof).