move unique track ID data migration into its own class
the data migration code to ensure that there are unique track IDs
doesn't really belong in the MediaFile class. This commit moves the
logic into its own class. This creates a pattern for data migrations
that allows the code handing the migration to be separate from the
database models that it modifies.