check edit media form validity before allowing changes to be saved
This commit adds client-side JavaScript code that checks the edit-media
form as the user is changing the form. It then performs a server-side
check that the provided values are valid. If they are not valid, the
client-side code updates the form to show the errors and disables the
submit button.

This stops the user from creating new media files that will fail the
validation checks.