add media handlers for multi-period streams
init segments and media segments need to be handled differently for
multi-period streams, as each Period defines where to start within
a stream. When live mps-streams is added, this will also benefit from
using this different media request handlers, as it will not need to
perform the same "how many times has the stream looped?" calculations.