combine manifest link and manifest edit input
when viewing the contents of a manifest file, there was both a anchor
showing the URL of the manifest and a text input element that allows
the URL to be modified. This commit combines them so that only one of
them is visible at a time. To the user it appears to be one component
that changes between being an anchor and an input.