created tokens are JWT, but only the JTI is used when checking tokens
each API request from the client must include either an access token
or a refresh token. This token is a JSON Web Token (JWT). However,
when tokens are being checked for validity in a flask_jwt_extended
hook, only the unique identifier (JTI) from the token is provided.

The Token class was not correctly checking the jti during the creation
process, and the type hints were not correct. They assumed that the
JWT and decoded JWT were the same.