refactor checking if refresh token is valid from whoAmI hook into ApiRequests
rather than use a hook to check if a refresh token is valid and to
discover the details of a logged in user, put this logic into the
ApiRequests class. This makes the whoAmI hook simpler and makes sure
that before a REST request is made that needs an access token, the
server is contacted to get a token.