wait for user check to complete before rendering child components
if the ProtectedPage component is used on a page where it is required
(i.e. optional===false) then it must not render the child components
until the REST request to check the user has completed. Otherwise it
will briefly render the child component when the request is started and
not yet completed.