add netifaces2 package for IP address discovery
To allow socket.io to configure its cross-origin allow list, it needs
to know both all the IP addresses and their hostnames. This is because
the flask_socketio library is performing a simple string comparison
to the cors_allowed_origins array. To allow a websock dev server running
the frontend code, this list also needs to include the port that is
running this server.