create and tag database when first starting Docker container
When the Docker container first starts, it creates the sqlite
database. The second restart of the container was when causing the
database to be tagged, so that Alembic knew its version. This can
cause a problem is the service is started once, and only ever
restarted following an upgrade. This causes the wrong db version to
be set. This can then cause migrations to fail as Alembic tries to
apply the wrong migrations.