support generation of streams with TTML text tracks
add a "--subtitles" option that can be used to specify a file containing
subtitle data. This data can be in any of the following formats:

* TTML
* SCC / CEA 608
* EBU STL
* SRT
* WebVTT

The output will always be in TTML IMSC format. The fragments will be
given a duration that is twice the duration of the audio and video
fragments.