create a Docker container to run media encoder
having to build and install ffmpeg, ffprobe and MP4Box on a local
machine can be a bit of a pain, given all of the libraries upon which
they depend. Having a Docker container that has a known working
combination of tools makes this media encoder scripts much easier to
use.