use tokens rather than username and password when uploading streams
It is undesirable to require the use of usernames and passwords on the
command line, as these are likely to become stored in command history.
Providing a temporary access token (which is a password equivalent) for
uploads provides a slightly better solution. Whilst these tokens can
still end up being stored in command history, they expire after an
hour.