don't use arrow functions or vi.fn as a mock constructor
When mocking a class, vitest v3 allowed a vi.fn or an arrow function
to be used as the thing that is executed when "new MockedClass()" is
called. However, vitest v4 does not allow this and requires a function
or class to be given as a mock for a class constructor.